Basic sequencer code to play a sequence in parallel.
Create a Sequence class to trigger a set of sequences and wait for them all
to complete before calling a callback.
TODO: Tidy up detection of whether sequence / individual things are complete
- should this be handled in the sequencer?
TODO: Make each type of 'animateable' thing conform to the same interface.
TODO: Figure out how to template the pointer to the CallbackHandler so we can
have different types being called.