Get the original animations working.
Implement the original sequence we had on the picture along with rainbow & chase.
Add FLASH anim to NeoPixel_Section
TODO: Parallel sequences are buggy when callback is called.
TODO: Consider always updating sequences even if they are not the current one - theatre chase & rainbow don't work.
TODO: Change the different configs so steps can be used to prolong the sequence for things like chase / rainbow
(check scanner - that needs reverse, but we don't use it currently)
TODO: Check addressing - we seem to be overwriting banners, etc.